Implement chrome snippets extension and react app
The extension can be found in the folder `extension` (this one should be loaded in the browser). The snippets extension consists in a simple gmail extension that allows you to choose a predefined snippet that when clicked to be shown in your email body draft.

When you click in the snippets button added by the extension, an iframe is loaded pointing to the "/templates" server page, showing the drafts available for you. We use Heroku to make our server available online for two reasons:

1) It's practical. Otherwise I need to start my localhost every time I wanna test my extension;
2) It provides us an HTTPS url. We load an iframe inside gmail and a HTTPS url is required.