Fix bug of sending snippet to the draft when clicking in edit/destroy snippet
The whole table line was clicable, what was causing a misunderstood sending the snippet to the draft even when we were deleting/editing the snippet instead.
We may find a better solution to handle that, probably writing our own table specifying where it's clicable or not. But for know I'll keep that way.